Updated Tasks service and controller to check if the user owns the tasks before doing CRUD operations

+++ src/tasks/tasks.service.ts
@@ -1,9 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Task } from './entity/tasks.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
+import { User } from 'src/user/entities/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -11,22 +17,41 @@ export class TasksService {
     @InjectRepository(Task) private taskRepository: Repository<Task>,
   ) {}
 
-  async createTask(task: CreateTaskDto): Promise<Task> {
+  async createTask(task: CreateTaskDto, user: User): Promise<Task> {
     // Create task database
-    const newTask = this.taskRepository.create({ ...task });
+    const newTask = this.taskRepository.create({ ...task, user });
 
     return this.taskRepository.save(newTask);
   }
 
-  async deleteTask(id: string): Promise<void> {
+  async deleteTask(id: string, user: User): Promise<void> {
+    const tasks = await this.taskRepository.findOne({
+      where: { id },
+      relations: ['user'],
+    });
+
+    if (tasks.user?.id !== user.id) {
+      throw new ForbiddenException('Not allowed to delete this task');
+    }
     // Delete task database
     await this.taskRepository.delete(id);
   }
 
-  async updateTask(id: string, UpdateTaskDto: UpdateTaskDto): Promise<Task> {
-    const task = await this.taskRepository.findOne({ where: { id } });
+  async updateTask(
+    id: string,
+    UpdateTaskDto: UpdateTaskDto,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.taskRepository.findOne({
+      where: { id },
+      relations: ['user'],
+    });
     if (!task) {
       throw new NotFoundException();
+    }
+
+    if (user.id !== task.user.id) {
+      throw new ForbiddenException('Not allowed to update this task');
     }
 
     Object.assign(task, UpdateTaskDto);
@@ -34,10 +59,15 @@ export class TasksService {
     return await this.taskRepository.save(task);
   }
 
-  async getTasks(): Promise<Task[]> {
+  async getTasks(user: User): Promise<Task[]> {
     // Get all user tasks
     // Posible to get using filters
     // Or pagination
-    return await this.taskRepository.find();
+
+    const tasks = await this.taskRepository.find({
+      where: { user: { id: user.id } },
+    });
+
+    return tasks;
   }
 }
